Parse the revlist after rebase in order to get the commits to attach
When the user invokes the file or attach commands with symbolic names
to specify the commits to attach, they probably expect the symbolic
names to be parsed after the rebase that is required to edit the
commit message.  This patch causes us to reparse the revlist after
rebasing when performing an add-url operation during the file and
attach commands.

Opting out of this behavior is always possible by using a commit SHA1.